Kotlin - Literal Constants Added.
